Move analytics and funnel files
