+++ src/frontend/src/flows/authorize/postMessageInterface.ts
@@ -1,6 +1,6 @@
 // Types and functions related to the window post message interface used by
 // applications that want to authenticate the user using Internet Identity
-import { analytics } from "$src/utils/analytics";
+import { analytics } from "$src/utils/analytics/analytics";
 import { type SignedDelegation as FrontendSignedDelegation } from "@dfinity/identity";
 import { Principal } from "@dfinity/principal";
 import { z } from "zod";
